add extra info

+++ src/http2.rs
@@ -53,7 +53,7 @@ pub async fn handle(
             "tcp_rtt_ms": tcp_rtt_ms,
             "tcp_rtt_dev_ms": tcp_rtt_dev_ms,
             "http2_rtt_ms": tls_rtt / 1000.0,
-            "http2_rtt_dev_ms": tls_rtt_dev / 1000.0,
+            "http2_rtt_dev_ms": tls_rtt_dev as f32 / 1000.0,
         })
         .to_string()
         .into(),
